BUG: Array indexing error in parameter error msg
In `pystan.model.StanModel.sampling`, users can request to store a subset of model parameters. The function contains a check that the requested parameters exist, but the error message contains `pars[unmatched]`, when `unmatched` is already an array of requested parameters which don't exist. This produces the error `IndexError: arrays used as indices must be of integer (or boolean) type`, which is difficult for users to decode.