Use -Os (like -O2) optimization for stanc (#252)
To better support clang, optimization was reduced to -O0 in
stan-dev/pystan#33. In the naive hope that clang may have fixed its bugs
in the intervening three years and the desire to reduce the binary size
of PyStan wheels, this commit turns on -Os optimization which is
essentially -O2 optimization absent those optimizations which tend to
increase file size. See also stan-dev/stan#2025.